[auto-battler] add wager to pvp, limit users to 1 battle
